+++ src/vecsync/chat/openai.py
@@ -13,9 +13,9 @@ class OpenAiChat:
         self.assistant_name = f"vecsync-{self.vector_store.store.name}"
         self.assistant_id = self._get_or_create_assistant()
 
-        self.thread_id = None if new_conversation else self._get_thead_id()
+        self.thread_id = None if new_conversation else self._get_thread_id()
 
-    def _get_thead_id(self) -> str | None:
+    def _get_thread_id(self) -> str | None:
         settings = Settings()
 
         match settings["openai_thread_id"]:
@@ -26,20 +26,22 @@ class OpenAiChat:
                 return x.value
 
     def _get_or_create_assistant(self):
-        existing_assistants = self.client.beta.assistants.list()
+        settings = Settings()
 
-        for assistant in existing_assistants:
-            if assistant.name == self.assistant_name:
-                print(f"✅ Assistant found: {assistant.id}")
-                return assistant.id
-
-        return self._create_assistant()
+        match settings["openai_assistant_id"]:
+            case SettingExists() as x:
+                print(f"✅ Assistant found: {x.value}")
+                return x.value
+            case _:
+                return self._create_assistant()
 
     def _create_assistant(self) -> str:
         instructions = """You are a helpful research assistant that can search through a large number
-        of journals and papers to find revelant information. It is very important that you
-        remain factual and cite information from the sources provided to you in the 
-        vector store. You are not allowed to make up information"""
+        of journals and papers to help answer the user questions. You have been given a file store which contains
+        the relevant documents the user is referencing. These documents should be your primary source of information.
+        You may only use external knowledge if it is helpful in clarifying questions. It is very important that you
+        remain factual and cite information from the sources provided to you in the file store. You are not allowed
+        to make up information."""
 
         assistant = self.client.beta.assistants.create(
             name=self.assistant_name,
@@ -55,11 +57,13 @@ class OpenAiChat:
 
         settings = Settings()
         del settings["openai_thread_id"]
+        del settings["openai_assistant_id"]
 
         print(f"🖥️ Assistant created: {assistant.name}")
         print(
             f"🔗 Assistant URL: https://platform.openai.com/assistants/{assistant.id}"
         )
+        settings["openai_assistant_id"] = assistant.id
         return assistant.id
 
     def chat(self, prompt: str) -> str: